got the stupid mistake, really a rookie error... wasted days on this... but the inimax still evaluating to different positions for the same state... will correct this soon

+++ src/Grid/Voronoi.h
@@ -45,7 +45,7 @@ public:
 	std::string outputOne();
 	std::string outputTwo();
 	virtual ~Voronoi();
-	int * calculate(int);
+	float * calculate(int);
 	inline void setWidth(int w){width = w;}
 
 };
